Add files via upload
NPR_post_DLC_V5_batch.m change the way the user to define the two ROIs for the objects. It asks user draws a line of one of the open field boarder as 0.45m. It changed to let the user to click to specify the centers of the two ROIs. For the ROI radius, it is pre-define as 0.04 m.